Move map parsing out from obs integration
